+++ src/main/java/me/drex/fafpatch/impl/block/BaseFactoryBlock.java
@@ -2,6 +2,8 @@ package me.drex.fafpatch.impl.block;
 
 import eu.pb4.factorytools.api.block.FactoryBlock;
 import eu.pb4.factorytools.api.virtualentity.BlockModel;
+import eu.pb4.polymer.blocks.api.BlockModelType;
+import eu.pb4.polymer.blocks.api.PolymerBlockResourceUtils;
 import eu.pb4.polymer.blocks.api.PolymerTexturedBlock;
 import eu.pb4.polymer.virtualentity.api.ElementHolder;
 import me.drex.fafpatch.impl.model.generic.BSMMParticleBlock;
@@ -21,6 +23,7 @@ import java.util.function.Function;
 
 public record BaseFactoryBlock(BlockState clientState, boolean tick, Function<BlockState, BlockModel> modelFunction) implements FactoryBlock, PolymerTexturedBlock, BSMMParticleBlock {
     public static final BaseFactoryBlock BARRIER = new BaseFactoryBlock(Blocks.BARRIER. defaultBlockState(), false, BlockStateModel::longRange);
+    public static final BaseFactoryBlock PLANT = new BaseFactoryBlock(PolymerBlockResourceUtils.requestEmpty(BlockModelType.PLANT_BLOCK),false, BlockStateModel::midRange);
 
     @Override
     public BlockState getPolymerBlockState(BlockState state, PacketContext context) {